requirements.txt
fastapi
uvicorn
pandas
numpy
streamlit
requests
google-genai
pydantic
